q4_sentiment seems OK.
changing regularization, learning rate, and iteration times doesn't
seems to make a big difference.

Maybe it is such a model of small capacity.

the Softmax loss at q3_sgd is so high that I still think sth is wrong

this homework is not so friendly, I don't know if I'm wrong. o(一︿一+)o